third_party/WebKit/LayoutTests/typedcssom/cssPositionValue.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

test(function() {
  assert_equals(new CSSPositionValue(new CSSUnitValue(50, 'px'),
    new CSSCalcValue({px: -10, em: -3.2, pt: 0})).toString(), '50px calc((-3.2em - 10px) + 0pt)');
  assert_equals(new CSSPositionValue(new CSSUnitValue(50, 'px'),
    new CSSUnitValue(2, 'em')).toString(), '50px 2em');
  assert_equals(new CSSPositionValue(new CSSCalcValue({px: -10, em: -3.2, pt: 0}),
    new CSSCalcValue({px: -10, em: 3.2})).toString(), 'calc((-3.2em - 10px) + 0pt) calc(3.2em - 10px)');
  assert_equals(new CSSPositionValue(new CSSCalcValue({px: -10, em: -3.2, pt: 0}),
    new CSSUnitValue(10, 'percent')).toString(), 'calc((-3.2em - 10px) + 0pt) 10%');
}, "toString() returns a string with the x and y positions cssStrings separated by a space");

</script>

<body>
</body>
